Purchase viewModel linked - PurchaseRepo.kt some parts are remaining

+++ app/src/main/java/com/vky342/openerp/data/Entities/Purcahases.kt
@@ -10,5 +10,5 @@ data class Purcahase(
     var purchaseId : Int,
     var purchaseDate : String,
     var ledgerId  : Int,
-    var purchaseAmount : Int,
+    var purchaseAmount : Double,
     )
